+++ nox/src/connectivity.rs
@@ -15,18 +15,16 @@
  */
 
 use std::cmp::min;
-use std::collections::{HashMap, HashSet};
-use std::sync::Arc;
+use std::collections::HashSet;
 use std::time::Duration;
 
+use crate::health::ConnectivityHealthChecks;
 use connection_pool::{ConnectionPoolApi, ConnectionPoolT, LifecycleEvent};
 use fluence_libp2p::PeerId;
 use futures::{stream::iter, StreamExt};
-use health::HealthCheck;
 use humantime_serde::re::humantime::format_duration as pretty;
 use kademlia::{KademliaApi, KademliaApiT, KademliaError};
 use libp2p::Multiaddr;
-use parking_lot::Mutex;
 use particle_protocol::{Contact, Particle, SendStatus};
 use peer_metrics::{ConnectivityMetrics, Resolution};
 use tokio::time::sleep;
@@ -277,51 +275,5 @@ impl AsRef<KademliaApi> for Connectivity {
 impl AsRef<ConnectionPoolApi> for Connectivity {
     fn as_ref(&self) -> &ConnectionPoolApi {
         &self.connection_pool
-    }
-}
-
-#[derive(Clone)]
-pub struct ConnectivityHealthChecks {
-    pub bootstrap_nodes: BootstrapNodesHealthCheck,
-}
-
-#[derive(Clone)]
-pub struct BootstrapNodesHealthCheck {
-    bootstrap_nodes_statuses: Arc<Mutex<HashMap<Multiaddr, bool>>>,
-}
-
-impl BootstrapNodesHealthCheck {
-    pub fn new(bootstrap_nodes: Vec<Multiaddr>) -> Self {
-        let bootstrap_nodes_statuses = bootstrap_nodes
-            .into_iter()
-            .map(|addr| (addr, false))
-            .collect::<HashMap<_, _>>();
-        Self {
-            bootstrap_nodes_statuses: Arc::new(Mutex::new(bootstrap_nodes_statuses)),
-        }
-    }
-
-    fn on_bootstrap_disconnected(&self, addresses: Vec<Multiaddr>) {
-        let mut guard = self.bootstrap_nodes_statuses.lock();
-        for addr in addresses {
-            guard.insert(addr, false);
-        }
-    }
-
-    fn on_bootstrap_connected(&self, addr: Multiaddr) {
-        let mut guard = self.bootstrap_nodes_statuses.lock();
-        guard.insert(addr, true);
-    }
-}
-
-impl HealthCheck for BootstrapNodesHealthCheck {
-    fn check(&self) -> eyre::Result<()> {
-        let guard = self.bootstrap_nodes_statuses.lock();
-        for (addr, connected) in guard.iter() {
-            if !connected {
-                return Err(eyre::eyre!("Bootstrap {} is not connected", addr));
-            }
-        }
-        Ok(())
     }
 }
